// models/dbTblNames.model.ts
export enum EDBTableTitles {
  ARTICLE = 'tbl_useful',
  ARTICLE_CATEGORIES = 'tbl_categories',
  TRANS_NEWS = 'tbl_digest',
  CHANNELS = 'tbl_channals',
  CHANNEL_SAT = 'tbl_chan_sat',
  CHANNEL_THEME = 'tbl_chan_tema',
  CHANNEL_ENCRYPTION = 'tbl_chan_encryption',
  CHANNEL_FREQUENCY = 'tbl_chan_freq',
  CHANNEL_BEAM = 'tbl_chan_beam',
  CHANNEL_COMPRESSION = 'tbl_chan_compress',
  TBL_LANGUAGE = 'tbl_language',
  CHANNEL_CATEGORY = 'tbl_chan_categ',
  COMMENTS_ARTICLE = 'tbl_comments',
  COMMENTS_CHANNEL = 'tbl_comments_chan',
  COMMENTS_CHAT = 'tbl_comments_chat',
  COMMENTS_INSTALLATION = 'tbl_comments_instal',
  COMMENTS_MAPS = 'tbl_comments_maps',
  COMMENTS_ONLINE = 'tbl_comments_online',
  COMMENTS_PACKAGES = 'tbl_comments_packs',
  COMMENTS_SATELLITE = 'tbl_comments_sat',
  COMMENTS_GENRE = 'tbl_comments_tema',
  TV_SCHEDULE_VIPIKO = 'tv_shedule_vipiko',
  VIPIKO_CHANNELS = 'vipiko_chan',
  TV_SCHEDULE_VSE_TV = 'tv_shedule_vsetv',
  TV_SCHEDULE = 'tv_shedule',
  FLY_SATELLITES = 'fly_satellites',
  FLY_CHANNELS = 'fly_channels',
  FLY_PACKAGES = 'fly_chan_packages',
}

export const getDbTableLink = (tblName: EDBTableTitles) =>
  `https://installsat.tv/tvefir/index.php?route=/sql&pos=0&db=installsat&table=${tblName}`;