// models/url/url.model.ts
export const MAIN_URL = 'https://installsat.tv';

export enum EUrlBaseParam {
  BASE_PATH = '/',
  LANG = 'lang',
  DATE = 'date',
  URL_DATE = 'url_date',
  SLUG = 'slug',
  SATELLITE = 'sat',
  ARTICLE_PARAM = 'article',
  CATEGORY = 'cat',
  TRANSPONDER_NEWS = 'transponderni-novyny',
  PACKAGE_CHANNEL_LIST = 'spysok-kanaliv-paketu',
  SAT_CHANNEL_LIST = 'spysok-kanaliv-suputnyka',
  INSTALLATION_OPTIONS = 'varianty-vstanovlennia-anten',
  ARTICLE = 'stattia',
  SAT_COVERAGE_MAP = 'karty-pokryttia-suputnykiv',
  SAT_FINDER = 'satellite-finder',
  NEWS_AND_ARTICLES = 'novyny-ta-statti',
  CHANNEL_PARAMS = 'parametry-kanalu',
  KANAL = 'kanal',
  ONLINE_CHANNEL_LIST = 'telekanaly-onlain',
  CHANNELS_TV_PROGRAM = 'programa-telekanaliv',
  PRODUCT = 'tovar',
  PRODUCT_LIST = 'spysok-tovariv',
  PRODUCT_CATEGORIES = 'kategoriji-tovariv',
  DELETE_COMMENT_SUBSCRIPTION = 'delete-subscription',
  SIGN_IN = 'login',
  SIGN_UP = 'login/sign-up',
}

export const URL_SEARCH_PARAM_VALUE_FALSE = 'off';